adding config starter default credentails, project id => now using autowired config from core starter TODO : tests
