--- spring-cloud-gcp-config/src/main/java/org/springframework/cloud/gcp/config/GoogleConfigEnvironment.java
@@ -1,35 +1,0 @@
-/*
- *  Copyright 2017 original author or authors.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.springframework.cloud.gcp.config;
-
-import java.util.List;
-
-final class GoogleConfigEnvironment {
-
-	private List<Variable> variables;
-
-	GoogleConfigEnvironment() {
-	}
-
-	List<Variable> getVariables() {
-		return this.variables;
-	}
-
-	void setVariables(List<Variable> variables) {
-		this.variables = variables;
-	}
-}
