+++ spring-cloud-gcp-examples/spring-cloud-gcp-config-example/src/main/java/com/example/HelloController.java
@@ -1,33 +1,39 @@
 /*
- * Copyright 2017 Google Inc.
+ *  Copyright 2017 original author or authors.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
  *       http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
+
 package com.example;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+/**
+* Sample REST Controller to demonstrate Google Cloud Configuration Integration.
+*
+* @author Jisha Abubaker
+*/
 @RestController
 public class HelloController {
 
-  @Autowired
-  SampleConfig sampleConfig;
+	@Autowired
+	SampleConfig sampleConfig;
 
-  @RequestMapping("/")
-  public String index() {
-    return String.valueOf("Queue size : " + sampleConfig.getQueueSize() +
-        "/" + "X Enabled : " + sampleConfig.isFeatureXEnabled());
-  }
+	@RequestMapping("/")
+	public String index() {
+		return String.valueOf("Queue size : " + this.sampleConfig.getQueueSize() +
+				"/" + "X Enabled : " + this.sampleConfig.isFeatureXEnabled());
+	}
 }
