+++ spring-cloud-gcp-examples/spring-cloud-gcp-config-example/src/main/java/com/example/SampleConfig.java
@@ -1,24 +1,29 @@
 /*
- * Copyright 2017 Google Inc.
+ *  Copyright 2017 original author or authors.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
  *       http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
+
 package com.example;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * Sample Configuration class with property values loaded using the Runtime Configurator API.
+ * @author Jisha Abubaker
+ */
 @RefreshScope
 @Configuration
 public class SampleConfig {
@@ -30,18 +35,18 @@ public class SampleConfig {
 	private boolean isFeatureXEnabled;
 
 	public int getQueueSize() {
-		return queueSize;
+		return this.queueSize;
 	}
 
 	public void setQueueSize(int size) {
-		this.queueSize = queueSize;
+		this.queueSize = size;
 	}
 
 	public boolean isFeatureXEnabled() {
-		return isFeatureXEnabled;
+		return this.isFeatureXEnabled;
 	}
 
 	public void setFeatureXEnabled(boolean featureXEnabled) {
-		isFeatureXEnabled = featureXEnabled;
+		this.isFeatureXEnabled = featureXEnabled;
 	}
 }
