--- spring-cloud-gcp-examples/spring-cloud-gcp-config-example/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.cloud.bootstrap.BootstrapConfiguration=org.springframework.cloud.gcp.config.GoogleConfigPropertySourceLocator
