+++ spring-cloud-gcp-examples/spring-cloud-gcp-config-example/src/main/resources/application.properties
@@ -1,20 +1,13 @@
-# Copyright 2017 Google Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-spring.application.name=myapp
+
+# Sample configuration
+
+# Provide project id, credentials to override default
+#spring.cloud.gcp.project-id=my-gcp-project-id
+#spring.cloud.gcp.credentials-location=my-credentials-file-location
 
 # spring.application.name will be used if `spring.cloud.config.name` is not provided
-#spring.cloud.config.name=myapp
+spring.application.name=myapp
+# spring.cloud.config.name=myapp
 
 spring.cloud.config.profile=prod
