+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-config/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,4 @@
+#org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+#org.springframework.cloud.gcp.core.autoconfig.GcpContextAutoConfiguration
+org.springframework.cloud.bootstrap.BootstrapConfiguration=\
+org.springframework.cloud.gcp.config.GoogleConfigPropertySourceLocator
